#5658: moving sharding def into a new file.
