+++ tt_eager/tt_dnn/kernels/dataflow/reader_unary_sharded.cpp
@@ -11,8 +11,5 @@ void kernel_main() {
     uint32_t num_tiles_per_core = get_arg_val<uint32_t>(0);
     constexpr uint32_t cb_id_in0 = get_compile_time_arg_val(0);
 
-    constexpr uint32_t onetile = 1;
-    for (uint32_t i = 0; i < num_tiles_per_core; ++ i) {
-        cb_push_back(cb_id_in0, onetile);
-    }
+    cb_push_back(cb_id_in0, num_tiles_per_core);
 }
