+++ tt_eager/tt_dnn/module.mk
@@ -22,6 +22,7 @@ TT_DNN_SRCS = \
 	tt_eager/tt_dnn/op_library/eltwise_unary/eltwise_unary_op.cpp \
 	tt_eager/tt_dnn/op_library/eltwise_unary/single_core/eltwise_unary_op_single_core.cpp \
 	tt_eager/tt_dnn/op_library/eltwise_unary/multi_core/eltwise_unary_op_multi_core.cpp \
+	tt_eager/tt_dnn/op_library/eltwise_unary/multi_core/eltwise_unary_op_multi_core_height_block_sharded.cpp \
 	tt_eager/tt_dnn/op_library/pad/pad_op.cpp \
 	tt_eager/tt_dnn/op_library/pad/pad_op_multi_core.cpp \
 	tt_eager/tt_dnn/op_library/unpad/single_core/unpad_op_single_core.cpp \
